+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file. This projec
 
 ## Unreleased
 
+### Added
+
+- Allow an outbound integration event handler to implement a `publish()` method. The `handle()` method is still
+  supported, but `publish()` makes more sense to describe what the handler does with the event it has been given.
+
 ## [1.0.0] - 2024-03-09
 
 ### Removed